requirements.txt
discord.py~=1.7.3
python-dotenv~=0.18.0
PyNaCl==1.3.0
dnspython==1.16.0
replit~=3.1.0